requirements.txt
Adafruit-Blinka==8.6.0
adafruit-circuitpython-adxl34x==1.12.6
adafruit-circuitpython-busdevice==5.2.3
adafruit-circuitpython-neopixel==6.3.6
adafruit-circuitpython-pixelbuf==1.1.8
adafruit-circuitpython-requests==1.12.8
adafruit-circuitpython-typing==1.8.2
Adafruit-PlatformDetect==3.32.0
Adafruit-PureIO==1.1.9
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cached-property==1.5.2
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
colorzero==1.1
dbus-python==1.3.2
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distro==1.5.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.4
executing==1.1.0
fastjsonschema==2.16.2
gpiozero==1.6.2
idna==2.10
importlib-metadata==5.0.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==3.2.0
jupyter-client==7.3.5
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
Mako==1.1.3
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
more-itertools==4.2.0
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.6.1
nest-asyncio==1.5.6
notebook==6.4.12
numpy==1.23.3
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
picamera==1.13
picamera2==0.3.3
pickleshare==0.7.5
pidng==4.0.9
piexif==1.1.3
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.21.0
pycparser==2.21
pyftdi==0.54.0
Pygments==2.7.1
PyGObject==3.42.2
pyparsing==3.0.9
pyrsistent==0.15.5
pyserial==3.5
python-apt==2.2.1
python-dateutil==2.8.2
python-prctl==1.7
pyusb==1.2.1
PyYAML==5.3.1
pyzmq==24.0.1
requests==2.25.1
rpi-ws281x==4.3.4
RPi.GPIO==0.7.1
SCons==4.0.1
Send2Trash==1.8.0
simplejpeg==1.6.5
six==1.16.0
soupsieve==2.3.2.post1
spidev==3.5
ssh-import-id==5.10
stack-data==0.5.1
sysv-ipc==1.1.0
terminado==0.16.0
texttable==1.6.3
tinycss2==1.1.1
tornado==6.2
traitlets==5.4.0
typing-extensions==4.4.0
ufw==0.36
urllib3==1.26.5
v4l2-python3==0.3.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
zipp==1.0.0
